jpa 게시물 등록 dto

+++ nonameWright-Client-APP/src/main/java/com/site/noname/study/repository/StudyRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.site.noname.study.model.StudyEntity;
 
-public interface StudyRepository extends JpaRepository<StudyEntity, Integer> {
+public interface StudyRepository extends JpaRepository<StudyEntity, Long> {
 
 }
